Move OCR validation into validation package
