+++ scinoephile/image/ocr/validation/char_cursor.py
