+++ scinoephile/image/ocr/validation/char_dims.py
