+++ scinoephile/image/ocr/validation/char_grp_dims.py
