+++ scinoephile/image/ocr/validation/char_pair_gaps.py
