+++ scinoephile/image/ocr/validation/gap_cursor.py
