+++ scinoephile/image/ocr/validation/validation_manager.py
